update: hopefully fixed network packets

+++ TimeSpeed/Framework/Models/Messages/Interfaces/ITimeStateReplyMessage.cs
@@ -2,6 +2,6 @@ namespace TimeSpeed.Framework.Models.Messages.Interfaces
 {
     internal interface ITimeStateReplyMessage : ITimeInfoMessage
     {
-        public int Timeout { get; init; }
+        public float Timeout { get; init; }
     }
 }
